GetAllEvictions implementation

+++ WlihaHackEviction/Models/Model.cs
@@ -39,9 +39,9 @@ namespace WlihaHackEviction.Models
         [DataMember]
         public int ZipCode { get; set; }
         [DataMember]
-        public float Latitude { get; set; }
+        public double Latitude { get; set; }
         [DataMember]
-        public float Longitude { get; set; }
+        public double Longitude { get; set; }
     }
 
     [DataContract]
